Implemented Luis' feature-request for direct Tumblr-uploading (i.e. without a daemon). Abstracted away output handling into plugin-like functions; removed uploadtotumblr.py daemon, and added a regular outputhandler for Tumblr instead
